+++ CHANGELOG.md
@@ -4,8 +4,12 @@
 
 Version 0.4.0
 
-- Added configuration for the default seperator
+* Added configuration for the default seperator
 
 Version 0.5.0 
 
-- Bug fixes
+* Bug fixes
+
+Version 0.6.0
+
+* Conversion is moved to Javascript from Original Python script